requirements.txt
Django==1.9
Pillow==3.0.0
gunicorn==19.4.1